refactor: moved entities package outside of internal
