+++ entities/event.go
