+++ internal/processors/filter/filter.go
@@ -18,7 +18,7 @@ package filter
 import (
 	"fmt"
 
-	"github.com/mia-platform/integration-connector-agent/internal/entities"
+	"github.com/mia-platform/integration-connector-agent/entities"
 
 	"github.com/google/cel-go/cel"
 	"github.com/google/cel-go/common/types"
