+++ internal/processors/filter/filter_test.go
@@ -18,7 +18,7 @@ package filter
 import (
 	"testing"
 
-	"github.com/mia-platform/integration-connector-agent/internal/entities"
+	"github.com/mia-platform/integration-connector-agent/entities"
 
 	"github.com/stretchr/testify/require"
 )
