+++ internal/processors/processor.go
@@ -19,15 +19,11 @@ import (
 	"context"
 	"fmt"
 
+	"github.com/mia-platform/integration-connector-agent/entities"
 	"github.com/mia-platform/integration-connector-agent/internal/config"
-	"github.com/mia-platform/integration-connector-agent/internal/entities"
 	"github.com/mia-platform/integration-connector-agent/internal/processors/filter"
 	"github.com/mia-platform/integration-connector-agent/internal/processors/mapper"
 )
-
-type Processor interface {
-	Process(data entities.PipelineEvent) (entities.PipelineEvent, error)
-}
 
 var (
 	ErrProcessorNotSupported = fmt.Errorf("processor not supported")
@@ -39,7 +35,7 @@ const (
 )
 
 type Processors struct {
-	processors []Processor
+	processors []entities.Processor
 }
 
 func (p *Processors) Process(_ context.Context, message entities.PipelineEvent) (entities.PipelineEvent, error) {
