+++ internal/sinks/sink.go
@@ -19,7 +19,7 @@ import (
 	"context"
 	"errors"
 
-	"github.com/mia-platform/integration-connector-agent/internal/entities"
+	"github.com/mia-platform/integration-connector-agent/entities"
 )
 
 var (
